refactor: add `Locale.hpp`
