+++ src/MeoAssistant/MeoAssistant.vcxproj
@@ -125,6 +125,7 @@
     <ClInclude Include="Utils\AsstRanges.hpp" />
     <ClInclude Include="Utils\AsstTypes.h" />
     <ClInclude Include="Utils\AsstUtils.hpp" />
+    <ClInclude Include="Utils\Locale.hpp" />
     <ClInclude Include="Utils\Demangle.hpp" />
     <ClInclude Include="Utils\Meta.hpp" />
     <ClInclude Include="Utils\Logger.hpp" />
