+++ src/MeoAssistant/MeoAssistant.vcxproj.filters
@@ -629,6 +629,9 @@
     <ClInclude Include="Utils\Logger.hpp">
       <Filter>源文件\Utils</Filter>
     </ClInclude>
+    <ClInclude Include="Utils\Locale.hpp">
+      <Filter>源文件\Utils</Filter>
+    </ClInclude>
     <ClInclude Include="Utils\SingletonHolder.hpp">
       <Filter>源文件\Utils</Filter>
     </ClInclude>
